backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.8.0
tensorflow==2.20.0
